fix: match retail save format and template loading
Retail parity checks (PyKotor against the live engine, and a GFF field-type
diff of our saves against a retail SAVEGAME.sav) found engine state that was
lost or corrupted between template, save and load:

- GFFObject wrote empty lists as offset 0xFFFFFFFF; spec readers (PyKotor)
  reject the file. Empty lists now point at a zero count.
- ActionParameter.toStruct never wrote Type, so every saved action parameter
  reloaded as a hole (3C-FD's ActionPlayAnimation lost all three params).
- ModuleCreature loaded BodyVariation into bodyBag, never loaded
  fortbonus/refbonus/willbonus, saved refbonus twice and never willbonus, never
  saved Hologram, and left maxForcePoints undefined for retail templates
  (which carry no MaxForcePoints).
- ModuleSound never loaded Continuous, and AudioEmitter repeated every emitter;
  a UTS that is neither looping nor continuous now plays once, as retail does.
- Save field types now match retail: CurrentWeather BYTE, StealthXPLoss DWORD
  (+ StealthXPEnabled, real values), sound Times/Volume/VolumeVrtn BYTE,
  trigger Type INT, waypoint LocalizedName CEXOLOCSTRING, PT_CHEAT_USED BYTE.

+++ src/managers/PartyManager.ts
@@ -361,7 +361,7 @@ export class PartyManager {
         availNPCSList.addChildStruct(availStruct);
       }
 
-      partytable.RootNode.addField(new GFFField(GFFDataType.INT, 'PT_CHEAT_USED')).setValue(0);
+      partytable.RootNode.addField(new GFFField(GFFDataType.BYTE, 'PT_CHEAT_USED')).setValue(0);
       partytable.RootNode.addField(new GFFField(GFFDataType.INT, 'PT_CONTROLLED_NP')).setValue( GameState.getCurrentPlayer() == GameState.PartyManager.Player ? -1 : GameState.PartyManager.party.indexOf(GameState.getCurrentPlayer()) );
       partytable.RootNode.addField(new GFFField(GFFDataType.LIST, 'PT_COST_MULT_LIS'));
